Refactor: split common.tsx into utils and rename into core.tsx
